requirements-app.txt
# This is used to create Postgres stored functions via the migrations process.
alembic_utils >= 0.8.2

Authlib >= 1.3.0
boto3 >= 1.34.48
botocore >= 1.34.48
cachelib >= 0.12.0
cachetools >= 5.3.2
celery[sqs] >= 5.3.6
click-datetime >= 0.2

# Datadog Tracer and Profiler
ddtrace == 1.18.2

# 23 February, 2024: eventlet uses dnspython, but upgrading past 2.4.2 causes DNS errors.
dnspython == 2.4.2

docopt >= 0.6.2
eventlet ~= 0.33.3

# This is for hardware based authentication schemes.  (a.k.a. 2FA with security keys)
# https://developers.yubico.com/WebAuthn/
fido2 >= 1.1.2

Flask >= 3.0.2
Flask-Bcrypt >= 1.0.1
Flask-Cors >= 4.0.0
Flask-JWT-Extended >= 4.6.0
flask-marshmallow ~= 1.2.0
Flask-Migrate >= 4.0.5
Flask-SQLAlchemy >= 3.1.1
git+https://github.com/department-of-veterans-affairs/notification-utils.git

# The production image defined by Dockerfile uses Gunicorn.  The development image defined by Dockerfile.local
# runs the application with the Flask development server by default.
gunicorn[eventlet] >= 21.2.0

iso8601 >= 2.1.0
itsdangerous >= 2.1.2
jsonschema >= 4.21.1
kombu >= 5.3.5
marshmallow ~= 3.21.1
marshmallow-sqlalchemy ~= 1.0.0
nanoid >= 2.0.0
notifications-python-client >= 9.0.0

# Sending an SMS message with v3 endpoints uses this package to test for a valid phone number.  The package
# is also a dependency of notification-utils.
phonenumbers ~= 8.12.57

psycopg2-binary >= 2.9.9
pwnedpasswords >= 2.0.0
PyJWT >= 2.8.0
# 4 Dec 2023: Ignoring security vulnerability 62123 in Makefile. It is fixed in >= 3.17
pypdf >= 4.0.2

python-dotenv >= 1.0.1
python-magic >= 0.4.27

# jsonschema requires this package to validate the date-time format but doesn't include it as a dependency.
# Validation fails silently without it.
# https://python-jsonschema.readthedocs.io/en/latest/validate/?highlight=date-time#validating-formats
rfc3339-validator >= 0.1.4

sentry-sdk[flask] >= 1.40.5
SQLAlchemy ~= 2.0.28
twilio >= 8.13.0
Unidecode >= 1.3.8
validatesns >= 0.1.1

# Flask depends on this package.  It's listed here as a top level dependency--intentionally with no version--because
# it's directly imported in v3 error handling code to reference the exception BadRequest.
Werkzeug
